requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
jinja2
python-dotenv
prometheus-fastapi-instrumentator
python-multipart
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.2